[enhance] Organize project folders
