--- Client/src/App.tsx
@@ -1,8 +1,0 @@
-import AppQs from "./Components/ApplicationQs/AppQs";
-function App() {
-  return (
-    <>
-    </>
-  );
-}
-export default App;
