+++ Client/src/Components/ApplicationDet.tsx
