+++ Client/src/Pages/App.tsx
@@ -1,0 +1,10 @@
+import AppQs from "../Components/ApplicationQs/AppQs";
+
+function App() {
+  return (
+    <>
+      <AppQs />
+    </>
+  );
+}
+export default App;
