+++ Client/src/assets/Icons/Bar.tsx
