+++ Client/src/assets/Icons/CircleIcon.tsx
@@ -2,7 +2,6 @@ const CircleIcon = () => (
     <svg width="32" height="32" viewBox="0 0 32 32" fill="none" xmlns="http://www.w3.org/2000/svg">
         <circle cx="16" cy="16" r="9" transform="matrix(1 0 0 -1 0 32)" stroke="#4B5563" />
     </svg>
-
 );
 
 export default CircleIcon;
