+++ Client/src/assets/Icons/CompareIcon.tsx
