+++ Client/src/assets/Icons/SearchIcon.tsx
