+++ Client/src/assets/Icons/SlideBarIcon.tsx
