+++ Client/src/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import App from './App.tsx'
+import App from './Pages/App.tsx'
 import './Styles/index.css'
 
 createRoot(document.getElementById('root')!).render(
